create object list widget
this is a chunk of contents widget extracted to be reused in find dialog
